pyproject.toml
[tool.poetry]
name = "synevarec"
version = "0.1.0"
description = ""
authors = ["Vladimir ProvaLove <vprovalov@ya.ru>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
jupyterlab = "^3.1.2"
numpy = "1.20.1"
sdv = "0.8.0"
pandas = "1.1.4"
scipy = "1.4.1"
scikit-surprise = "1.1.1"
seaborn = "0.11.1"
kaleido = "0.2.1"
tensorflow = "2.2.0"
deepctr-torch = "0.2.5"
dload = "0.6"
modules = {path = "./modules", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
